+++ src/com/sonymobile/hackathon/fruitsvsbrains/GamePosition.java
@@ -1,30 +1,30 @@
 package com.sonymobile.hackathon.fruitsvsbrains;
 
 public class GamePosition {
-	private int xPosition;
-	private int yPosition;
+	private float xPosition;
+	private float yPosition;
 
-	public GamePosition(int xPosition, int yPosition) {
+	public GamePosition(float xPosition, float yPosition) {
 		super();
 		this.xPosition = xPosition;
 		this.yPosition = yPosition;
 	}
 
-	public int getxPosition() {
+	public float getxPosition() {
 		return xPosition;
 	}
 
-	public int getyPosition() {
+	public float getyPosition() {
 		return yPosition;
 	}
 	
 	public int distanceTo(GamePosition gamePosition){
-		int xDiff =gamePosition.getxPosition()-xPosition;
-		int yDiff =gamePosition.getyPosition()-yPosition;
+		float xDiff =gamePosition.getxPosition()-xPosition;
+		float yDiff =gamePosition.getyPosition()-yPosition;
 		return (int) Math.round(Math.sqrt(xDiff*xDiff+yDiff*yDiff));
 	}
 	
-	public void move(int x, int y){
+	public void move(float x, float y){
 		xPosition+=x;
 		yPosition+=y;
 	}
